wallet: precomputed tables wasm
